requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-rosetta==0.10.0
idna==3.7
pillow==10.3.0
polib==1.2.0
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
